auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pydantic==2.5.3
